+++ argopy/data_fetchers/__init__.py
@@ -4,18 +4,16 @@ Most of these fetchers are meant to be used and discovered automatically by the 
 and by utilities functions list_available_data_src() and list_available_index_src()
 """
 
-from .erddap_refdata import Fetch_box as CTDRefDataFetcher
-from . import erddap_data
-from . import erddap_index
-from . import argovis_data
-from . import gdacftp_data
-from . import gdacftp_index
-
-__all__ = (
-    "erddap_data",
-    "erddap_index",
-    "argovis_data",
-    "gdacftp_data",
-    "gdacftp_index",
-    "CTDRefDataFetcher",
-)
+# from . import erddap_data
+# from . import erddap_index
+# from . import argovis_data
+# from . import gdacftp_data
+# from . import gdacftp_index
+#
+# __all__ = (
+#     "erddap_data",
+#     "erddap_index",
+#     "argovis_data",
+#     "gdacftp_data",
+#     "gdacftp_index",
+# )